refactor: remove MissingPersonalDetails
